+++ app/Http/Controllers/PenjualanController.php
@@ -8,8 +8,8 @@ use App\Barang;
 use App\Kategori;
 
 use App\Penjualan;
-use App\PenjualanDetail;
-use App\PenjualanLog;
+use App\PenjualanItem;
+use App\PenjualanBayar;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
@@ -20,7 +20,7 @@ class PenjualanController extends Controller
      * Data for Json Format (all)
      */
     public function penjualanJson(){
-        $list = Penjualan::with('penjualanDetail', 'penjualanDetail.barang', 'penjualanLog')->get()->map(function($data, $key) {
+        $list = Penjualan::with('penjualanItem', 'penjualanItem.barang', 'penjualanBayar')->get()->map(function($data, $key) {
             //Make formated invoice
             $data->invoice_url = str_replace('/', '-', $data->penjualan_invoice);
 
@@ -29,14 +29,14 @@ class PenjualanController extends Controller
             $barang = array();
             $total = 0;
             //Detail
-            foreach($data->penjualanDetail as $detail){
+            foreach($data->penjualanItem as $detail){
                 $barang[] = $detail->barang->barang_nama." (".$detail->jual_qty.")";
                 $total = $total + ( ($detail->harga_jual * $detail->jual_qty) - $detail->diskon );
             }
 
             $biayaLain = 0;
             $bayar = 0;
-            foreach($data->penjualanLog as $log){
+            foreach($data->penjualanBayar as $log){
                 $biayaLain = $biayaLain + $log->biaya_lain;
                 $bayar = $bayar + $log->bayar;
             }
@@ -49,8 +49,8 @@ class PenjualanController extends Controller
 
             //Remove unnecessary data
             unset($data->penjualan_detail);
-            unset($data->penjualanDetail);
-            unset($data->penjualanLog);
+            unset($data->penjualanItem);
+            unset($data->penjualanBayar);
             return $data;
         });
 
@@ -118,22 +118,19 @@ class PenjualanController extends Controller
 
         //Jika sukses simpan data penjualan
         if((bool)$storePenjualan){
-            $status .= "Penjualan OKE";
+            $status = "Penjualan Berhasil";
             $result = true;
 
-            //Insert Penjualan Detail
+            //Get an array from field Penjualan Detail
             $items = $request->barang_id;
             //Get Penjualan
             $penjualan = Penjualan::where('penjualan_invoice', $invoice)->get();
             $item_num = 1;
-            foreach($items as $item => $key){
-                //Cek Stok Barang
+            foreach($items as $item => $key){//Cek Stok Barang
                 $barang = Barang::where('id', $request->barang_id[$item])->get();//Ambil data barang terkait
-                if($barang[0]->barang_stokStatus == "Aktif"){
-                    //Jika Stok Aktif
-                    if($barang[0]->barang_stok >= $request->qty[$item]){
-                        //Jika stok lebih
-                        $storePenjualanDetail = PenjualanDetail::create([
+                if($barang[0]->barang_stokStatus == "Aktif"){ //Jika Stok Aktif
+                    if($barang[0]->barang_stok >= $request->qty[$item]){ //Jika stok lebih
+                        $storepenjualanItem = penjualanItem::create([
                             'penjualan_id' => $penjualan[0]->id,
                             'barang_id' => $request->barang_id[$item],
                             'harga_beli' => $request->harga_beli[$item],
@@ -142,21 +139,30 @@ class PenjualanController extends Controller
                             'diskon' => $request->diskon[$item],
                         ]);
 
-                        $status .= ", Penjualan Detail OKE";
+                        $status = "Item Penjualan Berhasil";
                         $error = false;
                         $result = true;
                     } else {
                         //Stok kurang
-                        $storePenjualanDetail = 0;
+                        $storepenjualanItem = 0;
 
-                        $status .= ", Penjualan Detail GAGAL, barang_id ".$barang[0]->barang_nama.", stok : ".$barang[0]->barang_stok." | qty : ".$request->qty[$item];
+                        $status = "(Baris item ke-".$item_num.") Kesalahan pada barang ".$barang[0]->barang_nama.", stok : ".$barang[0]->barang_stok." | qty : ".$request->qty[$item];
                         $error = "Gagal pada barang baris ke-".$item_num." (".$barang[0]->barang_nama.") karena sisa stok kurang (Stok : ".$barang[0]->barang_stok.", Permintaan QTY : ".$request->qty[$item].")";
                         $result = false;
                         //Delete All Data related to this transaction
-                        $status .= " | ".$this->destroy($penjualan[0]->id);
+                        $this->destroy($penjualan[0]->id);
+
+                        //Force stop process and send an error
+                        $message = [
+                            'result' => $result,
+                            'error' => $error,
+                            'invoice' => '',
+                            'message' => $status,
+                        ];
+                        return response()->json($message);
                     }
                 } else {
-                    $storePenjualanDetail = PenjualanDetail::create([
+                    $storepenjualanItem = penjualanItem::create([
                         'penjualan_id' => $penjualan[0]->id,
                         'barang_id' => $request->barang_id[$item],
                         'harga_beli' => $request->harga_beli[$item],
@@ -164,24 +170,24 @@ class PenjualanController extends Controller
                         'jual_qty' => $request->qty[$item],
                         'diskon' => $request->diskon[$item],
                     ]);
-                    $status .= ", Penjualan Detail OKE, tanpa stok. Barang_id ".$request->barang_id[$item];
+
+                    $status = "Item Penjualan Berhasil";
                     $error = false;
                     $result = true;
                 }
-
                 $item_num++;
             }
 
-            if((bool) $storePenjualanDetail){
+            if((bool) $storepenjualanItem){
                 //Insert Penjualan Log jika Penjualan Detail berhasil
-                $storePenjualanLog = PenjualanLog::create([
+                $storepenjualanBayar = penjualanBayar::create([
                     'penjualan_id' => $penjualan[0]->id,
                     'user_id' => Auth::user()->id,
                     'pembayaran_tgl' => $request->pembayaran_tgl,
                     'biaya_lain' => $request->biaya_lain,
                     'bayar' => $request->bayar,
                 ]);
-                $status .= ", Penjualan Log OKE";
+                $status = "Input transaksi Berhasil";
                 $error = false;
                 $result = true;
             }
@@ -190,7 +196,7 @@ class PenjualanController extends Controller
         }
 
         $message = [
-            'status' => $result,
+            'result' => $result,
             'error' => $error,
             'invoice' => str_replace('/', '-', $invoice),
             'message' => $status,
@@ -207,7 +213,7 @@ class PenjualanController extends Controller
     public function show($invoice)
     {
         $id = str_replace('-', '/', $invoice);
-        $penjualan = Penjualan::where('penjualan_invoice', $id)->with('toko', 'kostumer', 'penjualanDetail', 'penjualanDetail.barang', 'penjualanLog')->firstOrFail();
+        $penjualan = Penjualan::where('penjualan_invoice', $id)->with('toko', 'kostumer', 'penjualanItem', 'penjualanItem.barang', 'penjualanBayar')->firstOrFail();
 
         return view('staff.penjualan.invoice', compact('penjualan'));
         // return $penjualan;
@@ -222,7 +228,7 @@ class PenjualanController extends Controller
     public function edit($invoice)
     {
         $id = str_replace('-', '/', $invoice);
-        $penjualan = Penjualan::where('penjualan_invoice', $id)->with('toko', 'kostumer', 'penjualanDetail', 'penjualanDetail.barang', 'penjualanDetail.barang.kategori', 'penjualanLog')->firstOrFail();
+        $penjualan = Penjualan::where('penjualan_invoice', $id)->with('toko', 'kostumer', 'penjualanItem', 'penjualanItem.barang', 'penjualanItem.barang.kategori', 'penjualanBayar')->firstOrFail();
 
         // return response()->json($penjualan);
         return view('staff.penjualan.edit', compact('penjualan'));
@@ -239,6 +245,9 @@ class PenjualanController extends Controller
     {
         $this->validator($request->all())->validate();
         $pesan = "";
+        $status = "";
+        $result = "";
+        $isChanged = array();
 
         $penjualan = Penjualan::findOrFail($id);
         //Penjualan
@@ -247,20 +256,69 @@ class PenjualanController extends Controller
         $penjualan->penjualan_tgl = $request->penjualan_tgl;
         $penjualan->penjualan_detail = $request->penjualan_detail;
 
-        $isChanged = $penjualan->isDirty();//Check if anything change
-        //$status_penjualan = $penjualan->save();
-
-        if($isChanged){
-            $pesan = 'Successfully updated!';
-        }
+        $isChanged[] = $penjualan->isDirty();//Check if anything change
+        $status_penjualan = $penjualan->save();
 
         //Detail
+        $items = $request->barang_id;
+        $item_num = 1;
+        foreach($items as $item => $key){
+            //Cek Stok Barang
+            $barang = Barang::where('id', $request->barang_id[$item])->get();//Ambil data barang terkait
+            $penjualanItem = penjualanItem::where([
+                ['penjualan_id', $id],
+                ['barang_id', $request->barang_id[$item]]
+            ])->firstOrFail();//Ambil data penjualanItem terkait
+
+            if($barang[0]->barang_stokStatus == "Aktif"){ //Jika Stok Aktif
+                //Cek perubahan permintaan
+                $permintaanQty = $request->qty[$item] - $penjualanItem->jual_qty;
+
+                if($barang[0]->barang_stok >= $permintaanQty){ //Cek sisa stok dengan perubahan permintaan
+                    //Jika sisa stok lebih
+                    $penjualanItem->harga_jual = $request->harga_jual[$item];
+                    $penjualanItem->jual_qty = $request->qty[$item];
+                    $penjualanItem->diskon = $request->diskon[$item];
+
+                    $status = "Item Penjualan Berhasil. Permintaan baru : ".$permintaanQty;
+                    $error = false;
+                    $isChanged[] = $penjualanItem->isDirty();
+
+                    $result = (bool) $penjualanItem->save();
+                } else {
+                    //Sisa stok kurang
+                    $storepenjualanItem = 0;
+
+                    $status = "(Baris item ke-".$item_num.") Kesalahan pada barang ".$barang[0]->barang_nama.", sisa stok : ".$barang[0]->barang_stok." | tambahan permintaan : ".$permintaanQty;
+                    $error = "(Baris item ke-".$item_num.") Kesalahan pada barang ".$barang[0]->barang_nama.", sisa stok : ".$barang[0]->barang_stok." | tambahan permintaan : ".$permintaanQty;
+                    $result = false;
+
+                    $message = [
+                        'status' => $result,
+                        'error' => $error,
+                        'invoice' => '',
+                        'message' => $status,
+                    ];
+                    return response()->json($message);
+                }
+            } else {
+                $penjualanItem->harga_jual = $request->harga_jual[$item];
+                $penjualanItem->jual_qty = $request->qty[$item];
+                $penjualanItem->diskon = $request->diskon[$item];
+
+                $status = "Item Penjualan Berhasil, tanpa stok. Barang_id ".$request->barang_id[$item];
+                $error = false;
+                $isChanged[] = $penjualanItem->isDirty();
+                $result = (bool) $penjualanItem->save();
+            }
+            $item_num++;
+        }
 
         //Log
-        $penjualanLog = PenjualanLog::where('penjualan_id', $id)->get();
+        $penjualanBayar = penjualanBayar::where('penjualan_id', $id)->get();
         $biayaLain = 0;
         $bayar = 0;
-        foreach($penjualanLog as $log){
+        foreach($penjualanBayar as $log){
             $biayaLain = $biayaLain + $log->biaya_lain;
             $bayar = $bayar + $log->bayar;
         }
@@ -269,7 +327,7 @@ class PenjualanController extends Controller
         $perubahan_bayar = $request->bayar - $bayar;
 
         if($perubahan_biayaLain != 0 || $perubahan_bayar != 0){
-            $storePenjualanLog = PenjualanLog::create([
+            $storepenjualanBayar = penjualanBayar::create([
                 'penjualan_id' => $penjualan->id,
                 'user_id' => Auth::user()->id,
                 'pembayaran_tgl' => $request->pembayaran_tgl,
@@ -277,14 +335,20 @@ class PenjualanController extends Controller
                 'bayar' => $perubahan_bayar,
             ]);
 
-            $isChanged = true;
+            $isChanged[] = true;
+        }
+
+        if(in_array(true, $isChanged)){
             $pesan = 'Successfully updated!';
+        } else {
+            $pesan = 'Nothing changed!';
         }
 
         $message = [
             'invoice' => str_replace('/', '-', $penjualan->penjualan_invoice),
             'isChanged' => $isChanged,
             'message' => $pesan,
+            'status' => $status,
         ];
 
         return response()->json($message);
@@ -299,17 +363,17 @@ class PenjualanController extends Controller
     public function destroy($id)
     {
         //Delete data from Database
-        $data_penjualanLog = PenjualanLog::where('penjualan_id',$id);
-        $data_penjualanDetail = PenjualanDetail::where('penjualan_id',$id);
+        $data_penjualanBayar = penjualanBayar::where('penjualan_id',$id);
+        $data_penjualanItem = penjualanItem::where('penjualan_id',$id);
         $data_penjualan = Penjualan::findOrFail($id);
 
-        if($data_penjualanLog != null){
-            $hapus_log = $data_penjualanLog->delete();
+        if($data_penjualanBayar != null){
+            $hapus_log = $data_penjualanBayar->delete();
         } else {
             $hapus_log = "Not Found";
         }
-        if($data_penjualanDetail != null){
-            $hapus_detail = $data_penjualanDetail->delete();
+        if($data_penjualanItem != null){
+            $hapus_detail = $data_penjualanItem->delete();
         } else {
             $hapus_detail = "Not Found";
         }