+++ app/PenjualanBayar.php
@@ -4,12 +4,12 @@ namespace App;
 
 use Illuminate\Database\Eloquent\Model;
 
-class PenjualanDetail extends Model
+class PenjualanBayar extends Model
 {
-    protected $table = 'tbl_penjualan_detail';
+    protected $table = 'tbl_penjualan_bayar';
 
     protected $fillable = [
-        "penjualan_id", "barang_id", "harga_beli", "harga_jual", "jual_qty", "diskon"
+        "penjualan_id", "user_id", "pembayaran_tgl", "biaya_lain", "bayar",
     ];
 
     //Set relation with Penjualan
@@ -17,7 +17,7 @@ class PenjualanDetail extends Model
         return $this->belongsTo('App\Penjualan', 'penjualan_id');
     }
     //Set relation with Barang
-    public function barang(){
-        return $this->belongsTo('App\Barang', 'barang_id');
+    public function user(){
+        return $this->belongsTo('App\User', 'user_id');
     }
 }