+++ database/migrations/2018_11_07_155343_create_tbl_penjualan_item.php
@@ -4,7 +4,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateTblPenjualanLog extends Migration
+class CreateTblPenjualanItem extends Migration
 {
     /**
      * Run the migrations.
@@ -13,27 +13,28 @@ class CreateTblPenjualanLog extends Migration
      */
     public function up()
     {
-        Schema::create('tbl_penjualan_log', function (Blueprint $table) {
+        Schema::create('tbl_penjualan_item', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('penjualan_id')->unsigned(); //Relation dengan tbl_penjualan (id)
-            $table->integer('user_id')->unsigned(); //Relation dengan user (id)
-            $table->timestamp('pembayaran_tgl')->default(DB::raw('CURRENT_TIMESTAMP'));
-            $table->integer('biaya_lain');
-            $table->integer('bayar');
+            $table->integer('barang_id')->unsigned(); //Relation dengan tbl_barang (id)
+            $table->integer('harga_beli');
+            $table->integer('harga_jual');
+            $table->integer('jual_qty');
+            $table->integer('diskon')->default('0');
             $table->timestamps();
         });
 
         //Set foreign key
-        Schema::table('tbl_penjualan_log', function(Blueprint $table){
+        Schema::table('tbl_penjualan_item', function(Blueprint $table){
             $table->foreign('penjualan_id')
                 ->references('id')
                 ->on('tbl_penjualan')
                 ->onDelete('cascade')
                 ->onUpdate('cascade');
 
-            $table->foreign('user_id')
+            $table->foreign('barang_id')
                 ->references('id')
-                ->on('users')
+                ->on('tbl_barang')
                 ->onDelete('cascade')
                 ->onUpdate('cascade');
         });
@@ -47,13 +48,13 @@ class CreateTblPenjualanLog extends Migration
     public function down()
     {
         //Drop foreign
-        Schema::table('tbl_penjualan_log', function(Blueprint $table){
-            $table->dropForeign('tbl_penjualan_log_penjualan_id_foreign');
+        Schema::table('tbl_penjualan_item', function(Blueprint $table){
+            $table->dropForeign('tbl_penjualan_item_penjualan_id_foreign');
         });
         //Drop foreign
-        Schema::table('tbl_penjualan_log', function(Blueprint $table){
-            $table->dropForeign('tbl_penjualan_log_user_id_foreign');
+        Schema::table('tbl_penjualan_item', function(Blueprint $table){
+            $table->dropForeign('tbl_penjualan_item_barang_id_foreign');
         });
-        Schema::dropIfExists('tbl_penjualan_log');
+        Schema::dropIfExists('tbl_penjualan_item');
     }
 }