+++ database/migrations/2018_11_11_141527_create_trigger_stok_update_penjualan.php
@@ -4,7 +4,7 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
 
-class CreateTriggerRollbackPenjualan extends Migration
+class CreateTriggerStokUpdatePenjualan extends Migration
 {
     /**
      * Run the migrations.
@@ -14,13 +14,13 @@ class CreateTriggerRollbackPenjualan extends Migration
     public function up()
     {
         DB::unprepared('
-        CREATE TRIGGER rollback_penjualan AFTER DELETE ON `tbl_penjualan_item` FOR EACH ROW
+        CREATE TRIGGER stokUpdate_penjualan AFTER UPDATE ON `tbl_penjualan_item` FOR EACH ROW
             BEGIN
                 UPDATE
                 tbl_barang SET
-                    barang_stok = CASE WHEN barang_stokStatus = "Aktif" THEN barang_stok + old.jual_qty WHEN barang_stokStatus = "Tidak Aktif" THEN 0 END
+                    barang_stok = CASE WHEN barang_stokStatus = "Aktif" THEN barang_stok - (new.jual_qty - old.jual_qty) WHEN barang_stokStatus = "Tidak Aktif" THEN 0 END
                 WHERE
-                    id = old.barang_id;
+                    id = new.barang_id;
             END
         ');
     }
@@ -32,6 +32,6 @@ class CreateTriggerRollbackPenjualan extends Migration
      */
     public function down()
     {
-        DB::unprepared('DROP TRIGGER `rollback_penjualan`');
+        DB::unprepared('DROP TRIGGER `stokUpdate_penjualan`');
     }
 }